requirements.txt
streamlit==1.55.0
pandas==2.3.3
plotly==6.6.0
kaggle>=2.0.1,<3
python-dotenv>=1.0,<2
